And on to the meat of the problem.
Without a more specific definition of what the initial and update sets are,
I'm assuming that two datasets represent two partial, point-in-time snapshots
dumped from external publisher databases.  As such, I'm not assuming that the
three files represent the same three books in the same order.

The data include versions of books, and the titles imply that these are
subsequent editions of the same books, so I ran with that.

I introduced the new model, Edition, of which a logical Book may have many.  In
in interest of safety, I'm also assuming that the Aliases uniquely refer to
Editions rather than Books.

The process_data_file command now attempts to create Books and Editions
idempotently by making a fuzzy match given the aliases provided.  It assumes
that at least one identifier will remain constant across Editions, otherwise
it must create a new Book.  The fuzzy match happens in
Edition.most_likely_edition_by_aliases, where (for now), it finds an existing
edition with the most matching aliases to those provided.

Note: I chose not to address schema migration in this commit, but I would
recommend bringing in the excellent south library to manage that.  In order to
try my new version, just run the initial and update sets back-to-back on a
fresh database.